requirements.txt
pandas==2.2.1
openpyxl==3.1.2
numpy==1.26.4
customtkinter==5.2.2
darkdetect==0.8.0
matplotlib==3.8.3
Pillow==10.2.0
pyinstaller==6.4.0